Preliminary documentation
Added direct access to common transformations to Path to eliminate the need to talk about affine transformations in general. Still need to generate documentation diagrams to explain the transformations, but did create the TransformationDemo to display them visually.

Opted not to include a default getCentroid() method in `Path` because of the vagaries of how a Path may be defined and/or closed. Better to calculate the centroid based on knowledge than trying to build a universal algorithm.

Signed-off-by: Seth Battis <sbattis@gannacademy.org>